Add composable NativeQueryEditor with backward-compatible shim
